% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase_data.R
\docType{data}
\name{ase_data}
\alias{ase_data}
\title{ASEplot example dataset}
\format{
A list containing two elements:
\describe{
\item{element1}{ASE data frame output from ASET}
\item{element2}{Merged exon boundaries data frame output from ASET}
...
}
}
\source{
ASET: An end-to-end pipeline for quantification and visualization of allele specific expression (in peer-review)
}
\usage{
ase_data
}
\description{
A toy dataset used to demonstrate how to use ASEplot.
}
\keyword{datasets}
