% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase_data_in_gene.R
\name{ase_data_in_gene}
\alias{ase_data_in_gene}
\title{Subset ASE data by a gene name and create an interpolated exon coordinate file}
\usage{
ase_data_in_gene(ase_data, exons_all, gene_symbol)
}
\arguments{
\item{ase_data}{ASE data frame, generated from ASEprep}

\item{exons_all}{Exon coordinates, generated from ASEprep}

\item{gene_symbol}{Gene symbol}
}
\value{
A list that contains the subsetted ASE data frame for the given gene, and the interpolated exon coordinates for plotting
}
\description{
Subset ASE data by a gene name and create an interpolated exon coordinate file
}
\examples{
data('ase_data.test')
ase_df = ase_data$ase_df
exons = ase_data$union_exons_per_gene
ase_data_in_gene(ase_df, exons, 'RHOBTB3')
}
