% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ase_heatmap.R
\name{gene_ase_heatmap}
\alias{gene_ase_heatmap}
\title{Plot SNP-level ASE or parental allele frequency distribution as a heatmap for a given gene.}
\usage{
gene_ase_heatmap(ase_data, exons_all, gene_symbol, data_type)
}
\arguments{
\item{ase_data}{ASE data frame, generated from ASEprep.}

\item{exons_all}{Exon coordinates, generated from ASEprep}

\item{gene_symbol}{Gene symbol.}

\item{data_type}{Either 'ase' (|0.5-RefFreq|), 'pat-freq' or 'mat-freq'.}
}
\value{
A ggplot in which each column is a SNP site and each row is a sample.
}
\description{
Plot SNP-level ASE or parental allele frequency distribution as a heatmap for a given gene.
}
\examples{
data('ase_data.test')
ase_df = ase_data$ase_df
exons = ase_data$union_exons_per_gene
gene_ase_heatmap(ase_df, exons, "RHOBTB3", 'pat-freq')
}
