% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_poe_histogram.R
\name{gene_poe_histogram}
\alias{gene_poe_histogram}
\title{Plot gene-level parental allele frequency distribution as histogram.}
\usage{
gene_poe_histogram(
  ase_data,
  gene_symbol,
  data_type,
  binwidth = NULL,
  sample_name = NULL
)
}
\arguments{
\item{ase_data}{ASE data frame, generated from ASEprep.}

\item{gene_symbol}{Gene symbol.}

\item{data_type}{Either 'pat-freq' or 'mat-freq'.}

\item{binwidth}{Histogram binwidth. Defaults to 0.01.}

\item{sample_name}{If a sample name is provided, a vertical line will be drawn to indicate its position relative to the histogram.}
}
\value{
A ggplot.
}
\description{
Plot gene-level parental allele frequency distribution as histogram.
}
\examples{
data('ase_data.test')
ase_df = ase_data$ase_df
gene_poe_histogram(ase_df, 'RHOBTB3', 'pat-freq', '123882')
}
