% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_poe_ridge.R
\name{gene_poe_ridge}
\alias{gene_poe_ridge}
\title{Plot gene-level parental allele frequency distribution as a ridge plot for given genes.}
\usage{
gene_poe_ridge(ase_data, gene_symbols, data_type)
}
\arguments{
\item{ase_data}{ASE data frame, generated from ASEprep.}

\item{gene_symbols}{Gene symbols.}

\item{data_type}{Either 'pat-freq' or 'mat-freq'.}
}
\value{
A ggplot.
}
\description{
Plot gene-level parental allele frequency distribution as a ridge plot for given genes.
}
\examples{
data('ase_data.test')
ase_df = ase_data$ase_df
gene_poe_ridge(ase_df, c("MEG3", "PEG3", "RHOBTB3"), 'pat-freq')
import ggridges
}
