% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_symbol.R
\name{get_gene_symbol}
\alias{get_gene_symbol}
\title{Get gene symbols from the genes_exonic column.}
\usage{
get_gene_symbol(gene_id_name)
}
\arguments{
\item{ase_data}{ASE data frame, generated from ASEprep}
}
\value{
A vector with the gene symbols split from the genes_exonic column
}
\description{
Get gene symbols from the genes_exonic column.
}
\examples{
data('ase_data.test')
ase_df = ase_data$ase_df
ase_df$genes_exonic_symbol = sapply(ase_df$genes_exonic, get_gene_symbol)
}
