% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_gene_ase_boxplot.R
\name{snp_gene_ase_boxplot}
\alias{snp_gene_ase_boxplot}
\title{Plot SNP-level ASE, or parental allele frequency distribution as a boxplot for a given gene for all samples.}
\usage{
snp_gene_ase_boxplot(ase_data, gene_symbol, data_type)
}
\arguments{
\item{ase_data}{ASE data frame, generated from ASEprep.}

\item{gene_symbol}{Gene symbol.}

\item{data_type}{Either 'ase' (|0.5-RefFreq|), 'pat-freq' or 'mat-freq'.}
}
\value{
A ggplot.
}
\description{
Plot SNP-level ASE, or parental allele frequency distribution as a boxplot for a given gene for all samples.
}
\examples{
data('ase_data.test')
ase_df = ase_data$ase_df
snp_gene_ase_boxplot(ase_df, 'RHOBTB3', 'pat-freq')
}
