% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_location.R
\name{snp_location}
\alias{snp_location}
\title{Plot SNP positions relative to the given gene.}
\usage{
snp_location(
  ase_data,
  exons_all,
  gene_symbol,
  transcripts_display,
  assembly,
  sample_name = NULL
)
}
\arguments{
\item{ase_data}{ASE data frame, generated from ASEprep.}

\item{exons_all}{Exon coordinates, generated from ASEprep}

\item{gene_symbol}{Gene symbol.}

\item{transcripts_display}{Either 'split' (keeps all the transcripts displayed separately) or 'collapse' (collapses the transcripts into a single model).}

\item{assembly}{Either 'hg38' or 'mm10'.}

\item{sample_name}{If a sample name is given, the SNP positions from this sample will be indicated.}
}
\value{
A plot that shows the positions of SNPs in the top track (or two tracks, when a sample_name is given).
}
\description{
Plot SNP positions relative to the given gene.
}
\examples{
data('ase_data.test')
ase_df = ase_data$ase_df
exons = ase_data$union_exons_per_gene
snp_location(ase_df, exons, "RHOBTB3", 'collapse', 'hg38', '123882')
}
